+++ contrib/reports/explore-server/src/client/MyCodeMirrors.tsx
@@ -4,8 +4,7 @@ import * as ReactCodeMirror from "react-codemirror2";
 import * as ReactDOM from "react-dom";
 import * as Conclusions from "./Conclusions";
 import { Alert, Weakness } from "../shared/shared-types";
-import { getToolIcon, getToolOrderID } from "./util";
-import { rangeToArray } from "../../../../../src/util";
+import { getToolIcon, getToolOrderID, rangeToArray } from "./util";
 
 require("codemirror/mode/javascript/javascript");
