+++ contrib/reports/explore-server/src/client/util.tsx
@@ -45,6 +45,14 @@ export function groupAlertsByTool(alerts: Alert[], toolIDs: string[]) {
   }));
 }
 
+export function rangeToArray(start: number, end: number): number[] {
+  let result = [];
+  for (let i = start; i <= end; i++) {
+    result.push(i);
+  }
+  return result;
+}
+
 export function makeToolColumnKey(toolID: string) {
   return "tool-column-for-" + toolID;
 }
