+++ contrib/reports/explore-server/src/server/index.ts
@@ -13,7 +13,6 @@ import {
   CWEString
 } from "../../../../../src/persistent-types";
 import * as ClientTypes from "../shared/shared-types";
-import { rangeToArray } from "../../../../../src/util";
 
 function makePermalink(repository: string, commit: CommitID) {
   let githubRepoPattern = /^https:\/\/github.com\/([^/]+)\/([^/]+)\.git$/;
