+++ src/util.ts
@@ -364,14 +364,6 @@ export function spawnChildProcess(
   });
 }
 
-export function rangeToArray(start: number, end: number): number[] {
-  let result = [];
-  for (let i = start; i <= end; i++) {
-    result.push(i);
-  }
-  return result;
-}
-
 export function convertSarifLogToAlerts(log: sarif.Log): BCVEAlert[] {
   return log.runs.flatMap(run =>
     run.results.map((r: sarif.Result): BCVEAlert => {
